feat: Update BuresWassersteinFlowMatching with gradient option
This commit updates the `BuresWassersteinFlowMatching` class to include a new `gradient` option in the configuration. The default value is set to `'riemannian'`. This option allows for the selection of different gradient computation methods in the code. The code changes include conditionally updating the `mccann_derivative_jit` function and the `psd_project_jit` function based on the selected gradient method.