feat: Update BuresWassersteinFlowMatching with matched noise option
This commit updates the `BuresWassersteinFlowMatching` class to include a new `matched_noise` option in the configuration. The default value is set to `False`. When `matched_noise` is set to `True`, the `mini_batch_ot_mode` is automatically set to `False`. This change allows for the selection of different noise generation methods in the code. The code changes include conditionally updating the `mini_batch_ot_mode` based on the `matched_noise` value.